Refactor out bitbucket and github operations behind interface
