+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Helper files to isolate the code that deals with specific source code repos
+ * @author dyuen
+ */
+package io.dockstore.webservice.helpers;
